+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 - Add SCIP functions SCIPchgCoefLinear, SCIPaddCoefLinear and SCIPdelCoefLinear
 - Add SCIP function SCIPgetSolTime and wrapper getSolTime
 - Add convenience methods relax and getVarDict
+- Add SCIP functions hasPrimalRay, getPrimalRay, getPrimalRayVal
 ### Fixed
+- Fixed typo in documentation of chgRhs
 - Pricer plugin fundamental callbacks now raise an error if not implemented
 - Brachrule plugin fundamental callbacks now raise an error if not implemented
 - Fixed segmentation fault when accessing the Solution class directly